requirements.txt
alembic==0.9.6
apispec==5.1.0
apispec-webframeworks==0.5.2
Babel==2.5.1
blinker==1.4
boto==2.49.0
boto3==1.17.107
certifi==2017.7.27.1
chardet==3.0.4
click==7.1.2
dnspython==2.1.0
dominate==2.3.1
elasticsearch==6.8.2
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Moment==0.11.0
flask-paginate==0.8.1
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.11.2
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
micawber==0.5.4
#minio==7.0.4
PyJWT==1.5.3
PyMySQL==1.0.2
python-dateutil>=2.7.0
python-dotenv==0.15.0
python-editor==1.0.3
pyyaml==5.4.1
pytz==2017.2
redis==3.5.3
requests==2.25.1
rq==1.7.0
rq-dashboard==0.6.1
six==1.11.0
SQLAlchemy==1.3.22
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.1

# requirements for Heroku
#psycopg2==2.7.3.1
#gunicorn==19.7.1
